# expression: implement vectorized evaluation for builtinArithmeticIntDivideIntSig