--- a/expression/builtin_arithmetic_vec.go
+++ b/expression/builtin_arithmetic_vec.go
@@ -734,11 +734,113 @@ func (b *builtinArithmeticDivideRealSig) vecEvalReal(input *chunk.Chunk, result 
 }
 
 func (b *builtinArithmeticIntDivideIntSig) vectorized() bool {
-	return false
+	return true
 }
 
 func (b *builtinArithmeticIntDivideIntSig) vecEvalInt(input *chunk.Chunk, result *chunk.Column) error {
-	return errors.Errorf("not implemented")
+	n := input.NumRows()
+	lh, err := b.bufAllocator.get(types.ETInt, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(lh)
+
+	if err := b.args[0].VecEvalInt(b.ctx, input, lh); err != nil {
+		return err
+	}
+
+	if err := b.args[1].VecEvalInt(b.ctx, input, result); err != nil {
+		return err
+	}
+	result.MergeNulls(lh)
+
+	rh := result
+	lhi64s := lh.Int64s()
+	rhi64s := rh.Int64s()
+	resulti64s := result.Int64s()
+
+	isLHSUnsigned := mysql.HasUnsignedFlag(b.args[0].GetType().Flag)
+	isRHSUnsigned := mysql.HasUnsignedFlag(b.args[1].GetType().Flag)
+
+	switch {
+	case isLHSUnsigned && isRHSUnsigned:
+		err = b.divideUU(result, lhi64s, rhi64s, resulti64s)
+	case isLHSUnsigned && !isRHSUnsigned:
+		err = b.divideUS(result, lhi64s, rhi64s, resulti64s)
+	case !isLHSUnsigned && isRHSUnsigned:
+		err = b.divideSU(result, lhi64s, rhi64s, resulti64s)
+	case !isLHSUnsigned && !isRHSUnsigned:
+		err = b.divideSS(result, lhi64s, rhi64s, resulti64s)
+	}
+	return err
+}
+
+func (b *builtinArithmeticIntDivideIntSig) divideUU(result *chunk.Column, lhi64s, rhi64s, resulti64s []int64) error {
+	for i := 0; i < len(lhi64s); i++ {
+		if result.IsNull(i) {
+			continue
+		}
+		lh, rh := lhi64s[i], rhi64s[i]
+		if rh == 0 {
+			return handleDivisionByZeroError(b.ctx)
+		}
+		resulti64s[i] = int64(uint64(lh) / uint64(rh))
+	}
+	return nil
+}
+
+func (b *builtinArithmeticIntDivideIntSig) divideUS(result *chunk.Column, lhi64s, rhi64s, resulti64s []int64) error {
+	for i := 0; i < len(lhi64s); i++ {
+		if result.IsNull(i) {
+			continue
+		}
+		lh, rh := lhi64s[i], rhi64s[i]
+		if rh == 0 {
+			return handleDivisionByZeroError(b.ctx)
+		}
+		val, err := types.DivUintWithInt(uint64(lh), rh)
+		if err != nil {
+			return err
+		}
+		resulti64s[i] = int64(val)
+	}
+	return nil
+}
+
+func (b *builtinArithmeticIntDivideIntSig) divideSS(result *chunk.Column, lhi64s, rhi64s, resulti64s []int64) error {
+	for i := 0; i < len(lhi64s); i++ {
+		if result.IsNull(i) {
+			continue
+		}
+		lh, rh := lhi64s[i], rhi64s[i]
+		if rh == 0 {
+			return handleDivisionByZeroError(b.ctx)
+		}
+		val, err := types.DivInt64(lh, rh)
+		if err != nil {
+			return err
+		}
+		resulti64s[i] = val
+	}
+	return nil
+}
+
+func (b *builtinArithmeticIntDivideIntSig) divideSU(result *chunk.Column, lhi64s, rhi64s, resulti64s []int64) error {
+	for i := 0; i < len(lhi64s); i++ {
+		if result.IsNull(i) {
+			continue
+		}
+		lh, rh := lhi64s[i], rhi64s[i]
+		if rh == 0 {
+			return handleDivisionByZeroError(b.ctx)
+		}
+		val, err := types.DivIntWithUint(lh, uint64(rh))
+		if err != nil {
+			return err
+		}
+		resulti64s[i] = int64(val)
+	}
+	return nil
 }
 
 func (b *builtinArithmeticPlusIntSig) vectorized() bool {